setting up costar waypoint manager as a separate package
