+++ costar_bringup/launch/utilities.launch
@@ -4,23 +4,37 @@
   <arg name="start_instructor" default="true" doc="(bool) Launch user interface."/>
   <arg name="instructor_cases" default="" doc="(string) Limit allowed plugins to specific cases."/>
 
-    <!-- Librarian -->
-    <node pkg="librarian_core" type="core.py" name="librarian_core"/>
+  <!-- Librarian -->
+  <node pkg="librarian_core" type="core.py" name="librarian_core"/>
 
-    <!-- Predicator -->
-    <include file="$(find predicator_bringup)/launch/core.launch">
-    </include>
+  <!-- Waypoint Manager 2.0 -->
+  <node pkg="costar_waypoint_manager"
+    type="waypoint_manager_server.py"
+    name="costar_waypoints_manager"/>
 
-    <node pkg="point_cloud_recorder" type="point_cloud_recorder.py" name="point_cloud_recorder_service">
-      <param name="id" value="costar"/>
-    </node>
+  <!-- Predicator -->
+  <include file="$(find predicator_bringup)/launch/core.launch">
+  </include>
 
-    <!-- Force sensor -->
-    <node if="$(arg start_robotiq_force_sensor)" pkg="robotiq_force_torque_sensor" type="rq_sensor" name="robotiq_force_torque_sensor"/>
+  <!-- Point cloud recorder service -->
+  <node pkg="point_cloud_recorder"
+    type="point_cloud_recorder.py"
+    name="point_cloud_recorder_service">
 
-    <!-- User interface -->
-    <include if="$(arg start_instructor)" file="$(find instructor_core)/launch/instructor.launch">
-      <arg name="instructor_cases" value="$(arg instructor_cases)"/>
-    </include>
+    <param name="id" value="costar"/>
+  </node>
+
+  <!-- Force sensor -->
+  <node if="$(arg start_robotiq_force_sensor)"
+    pkg="robotiq_force_torque_sensor"
+    type="rq_sensor"
+    name="robotiq_force_torque_sensor"/>
+
+  <!-- User interface -->
+  <include if="$(arg start_instructor)"
+    file="$(find instructor_core)/launch/instructor.launch">
+
+    <arg name="instructor_cases" value="$(arg instructor_cases)"/>
+  </include>
 
 </launch>
