+++ costar_robot/costar_waypoint_manager/setup.py
@@ -1,0 +1,12 @@
+#!/usr/bin/env python
+
+from distutils.core import setup
+from catkin_pkg.python_setup import generate_distutils_setup
+
+d = generate_distutils_setup(
+    packages=['costar_waypoint_manager'],
+    package_dir={'': 'src'},
+)
+
+setup(**d)
+
